+++ backend/canisters/user_index/api/src/updates/set_moderation_flags.rs
@@ -1,0 +1,12 @@
+use candid::CandidType;
+use serde::{Deserialize, Serialize};
+
+#[derive(CandidType, Serialize, Deserialize, Debug)]
+pub struct Args {
+    pub moderation_flags_enabled: Option<u32>,
+}
+
+#[derive(CandidType, Serialize, Deserialize, Debug)]
+pub enum Response {
+    Success,
+}